+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added for_argument, attach_arguments and auto_adapt_to_methods to __all__ in `satella.coding`
 * added time_us to __all__ in `satella.time`
+* refactored `satella.cassandra.parallel_for`